Clean up item overlay code a bit
